+++ include/ltg/ltg.hpp
@@ -1,5 +1,6 @@
 ﻿#pragma once
 
+#include "ltg/protection_bootstrap.hpp"
 #include "ltg/tui_routes.hpp"
 #include "ltg/runtime_repair.hpp"
 #include "ltg/version.hpp"